live-a2a-server/pyproject.toml
[project]
name = "live-agents-a2a"
version = "0.1.0"
description = "Generate or modify images using A2A and CrewAI"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "crewai[tools]>=0.95.0",
    "google-genai>=1.9.0",
    "a2a-sdk>=0.3.0",
    "sse-starlette>=3.0.2",
    "scholarly>=1.7.11",
    "click>=8.2.1",
    "supabase>=1.0.0",
    "pandas>=2.0.3",
    "langchain>=0.1.0",
    "langchain_openai>=0.2.14",
    "langchain_experimental>=0.3.4",
    "tabulate>=0.9.0",
]
